// src/utils/types.ts
export type dropDownDataType ={
  navData:Array<{
    id:number
    title:string
    url?:string
    subtitle?:Array<{
      id:number
      title:string
      url:string
    }>
  }>
}